add: serialized struct for CveLog #5

+++ syslog-common/src/lib.rs
@@ -4,10 +4,10 @@
 #[derive(Clone, Copy)]
 pub struct SysCallLog {
     pub ts            : u64,
-    pub syscall       : u64,
-    pub pid           : u64,
+    pub syscall       : u32,
+    pub pid           : u32,
     pub pname_bytes   : [u8; 16]
 }
 
 #[cfg(feature = "user")]
-unsafe impl aya::Pod for SysCallLog{} // 
+unsafe impl aya::Pod for SysCallLog{}
